Updates ****
Introduced a Dictionary that would keep track of all the contacts added in the program and spreadsheet. If the user tries to add a user with the same first and last name as a contact already in the DB, it will show the user that is not going to happen.  ( a little exercise of method overloading )

Added a validation for the delete button. When the user clicks on the delete button it will ask him if he is sure he/she wants to delete the entry/contact.